requirements.txt
certifi==2019.6.16
dj-database-url==0.5.0
Django==2.2.4
djangorestframework==3.10.2
gunicorn==19.9.0
psycopg2==2.8.3
pytz==2019.1
sqlparse==0.3.0
whitenoise==4.1.3
wincertstore==0.2
